requirements.txt
amqp
arrow
asgiref
attrs
beautifulsoup4
boto3
botocore
billiard
Brotli
celery
certifi
cffi
channels
charset-normalizer
click
click-didyoumean
click-plugins
click-repl
colorama
coverage
cryptography
cssselect2
datetime-truncate
defusedxml
diff-match-patch
dj-database-url
Django
django-admin-charts
django-allauth
django-auto-logout
django-bootstrap-datepicker-plus
django-bootstrap4
django-bower
django-braces
django-ckeditor
django-ckeditor-5
django-contrib-comments
django-cors-headers
django-countries
django-crispy-forms
django-debug-toolbar
django-defender
django-dotenv
django-environ
django-filter
django-import-export
django-jazzmin
django-js-asset
django-mailgun-provider
django-mathfilters
django-memoize
django-model-utils
django-multiselectfield
django-nvd3
django-paypal
django-phone-field
django-social-share
django-static-fontawesome
django-stubs
django-stubs-ext
django-taggit
django-templated-mail
django-tinymce
django-utils-six
djangorestframework
djangorestframework-simplejwt
docutils
et-xmlfile
flower
fonttools
gunicorn
html5lib
humanize
ics
idna
Jinja2
jmespath
kombu
MarkupPy
MarkupSafe
mypy
mypy-extensions
mysqlclient
oauthlib
odfpy
openpyxl
packaging
Pillow
prometheus-client
prompt-toolkit
psycopg2
pycparser
pydantic
pydyf
PyJWT
pyphen
python-dateutil
python-decouple
python-nvd3
python-slugify
python3-openid
pytz
PyYAML
redis
regex
reportlab
requests
requests-oauthlib
rjsmin
s3transfer
shortuuid
six
soupsieve
sqlparse
stripe
swapper
tablib
TatSu
tinycss2
tomli
tornado
types-pytz
types-PyYAML
typing_extensions
tzdata
Unidecode
urllib3
validate-email
vine
wcwidth
weasyprint
webencodings
websocket-client
whitenoise
xlrd
xlwt
zopfli
